backend/requirements.txt
flask
torch
torchvision
numpy
opencv-python
Pillow
matplotlib
flask_cors
